created graphql folder
