+++ package.json
@@ -14,7 +14,9 @@
     "apollo-server-express": "^2.9.12",
     "cors": "^2.8.5",
     "ejs": "^2.5.6",
-    "express": "^4.15.2"
+    "express": "^4.15.2",
+    "passport": "^0.4.0",
+    "passport-http-bearer": "^1.0.1"
   },
   "devDependencies": {
     "request": "^2.81.0",
