+++ src/graphql/resolvers/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Query: {
+    books: () => console.log('books');
+  }
+}
